requirements.txt
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
googleapis-common-protos==1.60.0
grpclib==0.4.6
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
Pillow==10.0.1
protobuf==4.24.3
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.0.6
viam-sdk==0.7.0
yarl==1.9.2